#47 - WIP - Basic formating for italic and bold and gui refactoring to extract htext_editor.py from main_window.py
