+++ src/const/fonts.py
@@ -8,6 +8,7 @@ FA_SOLID = os.path.join(FONT_AWESOME_DIRECTORY, "Font Awesome 6 Free-Solid-900.o
 
 HELVETICA_FONT_NAME = "Helvetica"
 BOLD_FONT = "bold"
+ITALIC_FONT = "italic"
 TEXT_SIZE_SECTION_HEADER = 12
 TEXT_SIZE_MENU = 9
 TEXT_SIZE_BOTTOM_BAR = 10
