+++ src/const/tags.py
@@ -16,4 +16,12 @@ LONG_SENTENCE_TAG_NAME_MID = "long_sentence_mid"
 LONG_SENTENCE_TAG_NAME_HIGH = "long_sentence_high"
 SEARCH_RESULT_TAG_NAME = "search_result"
 CURRENT_SEARCH_RESULT_TAG_NAME = "current_search_result"
-FORMATTING_TAGS = {PARAGRAPH_TAG_NAME, SEARCH_RESULT_TAG_NAME, CURRENT_SEARCH_RESULT_TAG_NAME}
+BOLD_TAG_NAME = "bold"
+ITALIC_TAG_NAME = "italic"
+FORMATTING_TAGS = {
+    PARAGRAPH_TAG_NAME,
+    SEARCH_RESULT_TAG_NAME,
+    CURRENT_SEARCH_RESULT_TAG_NAME,
+    BOLD_TAG_NAME,
+    ITALIC_TAG_NAME
+}
