+++ src/const/values.py
@@ -13,3 +13,5 @@ NLP_BATCH_SIZE = 8000
 VERSION = "1.1.0"
 GITHUB_REPO = "MartinHlavna/hector"
 CURRENT_PROJECT_VERSION = 1
+# A4 SIZE IN INCHES. WE LATER USE DPI TO SET EDITOR WIDTH
+A4_SIZE_INCHES = 8.27
